V13: Controller refactor, Front-end updates
