+++ models/readings.json
@@ -120,6 +120,56 @@
       "time": "29/08/2023, 09:08:48",
       "_id": "9351e948-c485-4b18-9217-8084a79e969c",
       "stationid": "944b68d7-812d-48e1-96f1-096f4ba0d999"
+    },
+    {
+      "code": 804,
+      "temperature": 26.93,
+      "windSpeed": 1.8,
+      "windDirection": 172,
+      "pressure": 1014,
+      "time": "31/08/2023, 09:42:37",
+      "_id": "beba33c7-ced0-4c22-9835-f2be153729fb",
+      "stationid": "3d730470-49de-4c54-8441-48efce0616a7"
+    },
+    {
+      "code": 803,
+      "temperature": 8.02,
+      "windSpeed": 7.2,
+      "windDirection": 240,
+      "pressure": 1006,
+      "time": "31/08/2023, 09:42:58",
+      "_id": "ecd5e1ef-d9b2-4a1d-b364-c6d53a4a67f2",
+      "stationid": "fbd0fa8a-7c9f-442f-b434-7bb9d821caec"
+    },
+    {
+      "code": 800,
+      "temperature": 20.52,
+      "windSpeed": 2.16,
+      "windDirection": 294,
+      "pressure": 1011,
+      "time": "31/08/2023, 09:43:23",
+      "_id": "ee024374-ad07-4f60-912f-4ade70eb9b44",
+      "stationid": "944b68d7-812d-48e1-96f1-096f4ba0d999"
+    },
+    {
+      "code": 802,
+      "temperature": 22.9,
+      "windSpeed": 0.78,
+      "windDirection": 324,
+      "pressure": 1011,
+      "time": "02/09/2023, 09:46:40",
+      "_id": "9856b2c4-6a96-4753-831f-2d7cb38b725c",
+      "stationid": "944b68d7-812d-48e1-96f1-096f4ba0d999"
+    },
+    {
+      "code": 803,
+      "temperature": 7.02,
+      "windSpeed": 4.12,
+      "windDirection": 190,
+      "pressure": 1019,
+      "time": "02/09/2023, 09:47:53",
+      "_id": "01b88800-225a-478a-9d93-fe91c8ed1952",
+      "stationid": "fbd0fa8a-7c9f-442f-b434-7bb9d821caec"
     }
   ]
 }
