+++ models/stations.json
@@ -7,28 +7,28 @@
       "longitude": 125.8198,
       "userid": "2d089554-701f-4bdb-815c-2a8c9d98885a",
       "latestReading": {
-        "weather": "Cloudy",
-        "weatherIcon": "solar:cloud-line-duotone",
-        "temperature": 21.61,
-        "temperatureFahrenheit": 70.9,
-        "windSpeed": 3.51,
-        "windDirection": 237,
-        "windDirectionCompass": "WSW",
-        "windSpeedBeaufort": 1,
-        "pressure": 1005,
-        "windChill": 23.13,
-        "time": "29/08/2023, 09:08:48",
-        "_id": "44faeac8-6efd-4036-b26c-e258620baf89",
+        "weather": "Partial Clouds",
+        "weatherIcon": "ph:cloud-sun",
+        "temperature": 22.9,
+        "temperatureFahrenheit": 73.22,
+        "windSpeed": 0.78,
+        "windDirection": 324,
+        "windDirectionCompass": "NW",
+        "windSpeedBeaufort": 0,
+        "pressure": 1011,
+        "windChill": 25.15,
+        "time": "02/09/2023, 09:46:40",
+        "_id": "a3805591-76e4-49c0-883e-f4c00f2fa7c6",
         "stationid": "944b68d7-812d-48e1-96f1-096f4ba0d999"
       },
       "_id": "944b68d7-812d-48e1-96f1-096f4ba0d999",
-      "tempTrend": "icon-park-twotone:up-two",
-      "windTrend": "",
-      "pressureTrend": "icon-park-twotone:down-two",
-      "maxTemp": 21.61,
+      "tempTrend": "",
+      "windTrend": "icon-park-twotone:down-two",
+      "pressureTrend": "",
+      "maxTemp": 22.9,
       "minTemp": 18.23,
       "maxWindSpeed": 3.51,
-      "minWindSpeed": 1.94,
+      "minWindSpeed": 0.78,
       "maxPressure": 1011,
       "minPressure": 1005
     },
@@ -38,29 +38,29 @@
       "longitude": 15.63,
       "userid": "2d089554-701f-4bdb-815c-2a8c9d98885a",
       "latestReading": {
-        "weather": "Partial Clouds",
-        "weatherIcon": "ph:cloud-sun",
-        "temperature": 9.02,
-        "temperatureFahrenheit": 48.24,
-        "windSpeed": 2.57,
-        "windDirection": 150,
-        "windDirectionCompass": "SSE",
+        "weather": "Cloudy",
+        "weatherIcon": "solar:cloud-line-duotone",
+        "temperature": 7.02,
+        "temperatureFahrenheit": 44.64,
+        "windSpeed": 4.12,
+        "windDirection": 190,
+        "windDirectionCompass": "S",
         "windSpeedBeaufort": 1,
-        "pressure": 1006,
-        "windChill": 9.66,
-        "time": "29/08/2023, 09:08:11",
-        "_id": "bd3f468e-f635-4432-9e5f-afb1e1245453",
+        "pressure": 1019,
+        "windChill": 6.71,
+        "time": "02/09/2023, 09:47:53",
+        "_id": "21bc5961-dd59-45c8-8e3f-36b5df8f7db1",
         "stationid": "fbd0fa8a-7c9f-442f-b434-7bb9d821caec"
       },
       "_id": "fbd0fa8a-7c9f-442f-b434-7bb9d821caec",
-      "tempTrend": "icon-park-twotone:up-two",
-      "windTrend": "icon-park-twotone:down-two",
+      "tempTrend": "icon-park-twotone:down-two",
+      "windTrend": "",
       "pressureTrend": "",
       "maxTemp": 9.02,
       "minTemp": 7.02,
-      "maxWindSpeed": 4.63,
+      "maxWindSpeed": 7.2,
       "minWindSpeed": 2.57,
-      "maxPressure": 1012,
+      "maxPressure": 1019,
       "minPressure": 1006
     },
     {
@@ -69,29 +69,29 @@
       "longitude": 132.46,
       "userid": "2d089554-701f-4bdb-815c-2a8c9d98885a",
       "latestReading": {
-        "weather": "Thunder",
-        "weatherIcon": "mdi:weather-thunder",
-        "temperature": 31.89,
-        "temperatureFahrenheit": 89.4,
-        "windSpeed": 1.71,
-        "windDirection": 178,
+        "weather": "Cloudy",
+        "weatherIcon": "solar:cloud-line-duotone",
+        "temperature": 26.93,
+        "temperatureFahrenheit": 80.47,
+        "windSpeed": 1.8,
+        "windDirection": 172,
         "windDirectionCompass": "S",
         "windSpeedBeaufort": 1,
-        "pressure": 1011,
-        "windChill": 34.33,
-        "time": "29/08/2023, 09:07:45",
-        "_id": "1bc3d5d3-888d-4998-94b3-e1c755200545",
+        "pressure": 1014,
+        "windChill": 29.1,
+        "time": "31/08/2023, 09:42:37",
+        "_id": "9225eeef-32ad-483e-8682-09180d661c31",
         "stationid": "3d730470-49de-4c54-8441-48efce0616a7"
       },
       "_id": "3d730470-49de-4c54-8441-48efce0616a7",
-      "tempTrend": "icon-park-twotone:up-two",
-      "windTrend": "",
+      "tempTrend": "",
+      "windTrend": "icon-park-twotone:up-two",
       "pressureTrend": "icon-park-twotone:up-two",
       "maxTemp": 31.89,
       "minTemp": 26.89,
-      "maxWindSpeed": 1.71,
+      "maxWindSpeed": 1.8,
       "minWindSpeed": 0.25,
-      "maxPressure": 1011,
+      "maxPressure": 1014,
       "minPressure": 1008
     }
   ]
