+++ public/style.css
@@ -1,3 +1,6 @@
+.navbar {
+     border-radius: 25px;
+}
 
 .box {
   height: 100%;
