+++ utils/station-analytics.js
@@ -23,14 +23,14 @@ export const stationAnalytics = {
     const latestReading = {      
       weather: latestWeather.weather,
       weatherIcon: latestWeather.icon,
-      temperature:  await dataConversions.rounder(lastReading.temperature),
-      temperatureFahrenheit: await dataConversions.rounder(temperatureFahrenheit),
-      windSpeed:  await dataConversions.rounder(lastReading.windSpeed),
-      windDirection:  await dataConversions.rounder(lastReading.windDirection),
+      temperature:  lastReading.temperature,
+      temperatureFahrenheit: temperatureFahrenheit,
+      windSpeed:  lastReading.windSpeed,
+      windDirection:  lastReading.windDirection,
       windDirectionCompass: windDirectionCompass,
       windSpeedBeaufort: windSpeedBeaufort,  
-      pressure:  await dataConversions.rounder(lastReading.pressure),
-      windChill: await dataConversions.rounder(windChill),
+      pressure:  lastReading.pressure,
+      windChill: windChill,
       time: lastReading.time,
       _id: v4(),
       stationid: station._id,
