+++ utils/trends.js
@@ -1,7 +1,31 @@
 import { stationStore } from "../models/station-store.js";
-
+import { dataConversions } from "./conversions.js";
+import fetch from 'node-fetch';
 
 export const dataTrends = {
+  
+    async checkForecast (station) {
+    const lat = station.latitude;
+    const long = station.longitude;
+    const apiKey = "0c109ad6bb8a0b5d8a284ce6061f12c6";
+    const url = `https://api.openweathermap.org/data/2.5/forecast?lat=` + lat + `&lon=` + long + `&appid=` + apiKey;
+    const search = await fetch(url);
+    const data = await search.json();
+    let trends = {
+      tempTrend: [],
+      trendLabels: [],
+    };
+    let j = 0;
+    for (let i = 4; i < data.list.length; i += 8) {
+      let tempTrend = await dataConversions.rounder((data.list[i].main.temp)-273.15);
+      trends.tempTrend.push(tempTrend);
+      const date = new Date(data.list[i].dt * 1000);
+      trends.trendLabels.push(`${date.getDate()}/${date.getMonth()}/${date.getFullYear()}`);
+      j++;
+    }
+    return trends;
+  },
+  
   async checkTrends(stationid) {
     const station = await stationStore.getStationByIdWithReadings(stationid);
     const upTrend = "icon-park-twotone:up-two";
@@ -40,7 +64,7 @@ export const dataTrends = {
     else if ((readings[0].pressure > readings[1].pressure) && (readings[1].pressure > readings[2].pressure))
       results.pressureTrend = downTrend;
     
-    return results
+    return results;
   }
   
 }
