+++ views/partials/add-reading.hbs
@@ -1,51 +1,52 @@
+
 {{#each station}}
-  <form class="box" action="/station/{{_id}}/addreading" method="POST">
-    <label class="label has-text-centered"><b>Add New Reading</b></label>
-    <div class="field is-horizontal">
-      <div class="field-body">
-        <div class="field">
-          <div class="control">
-            <div class="select">
-              <select class="input" name="code">
-                <option value="100">Clear</option>
-                <option value="200">Partial Clouds</option>
-                <option value="300">Cloudy</option>
-                <option value="400">Light Rain</option>
-                <option value="500">Heavy Shower</option>
-                <option value="600">Rain</option>
-                <option value="700">Snow</option>
-                <option value="800">Thunder</option>
-              </select>
+  <label class="label has-text-centered"><b>Add New Reading</b></label>
+  <div class = "columns is-full">
+    <form class="column is-2" action="/station/{{_id}}/autogeneratereading" method="POST">
+      <label class="label "><b>Automatic:</b></label>
+      <div class="field">   
+        <button class="button is-danger is-centered">
+          Add Auto Reading
+        </button>
+      </div>
+    </form>
+    <form class="column is-10" action="/station/{{_id}}/addreading" method="POST">
+      <label class="label"><b>Manual Input:</b></label>
+      <div class="field is-horizontal">
+        <div class="field-body">
+          <div class="field">
+            <div class="control">
+              <div class="select">
+                <select class="input" name="code">
+                  <option value="100">Clear</option>
+                  <option value="200">Partial Clouds</option>
+                  <option value="300">Cloudy</option>
+                  <option value="400">Light Rain</option>
+                  <option value="500">Heavy Shower</option>
+                  <option value="600">Rain</option>
+                  <option value="700">Snow</option>
+                  <option value="800">Thunder</option>
+                </select>
+              </div>
             </div>
           </div>
-        </div>
-        <div class="field">
-          <input class="input" type="text" placeholder="Enter Temperature" name="temperature">
-        </div>
-        <div class="field">
-          <input class="input" type="text" placeholder="Enter Wind Speed" name="windspeed">
-        </div>
-        <div class="field">
-          <input class="input" type="text" placeholder="Enter Wind Direction" name="winddirection">
-        </div>
-        <div class="field">
-          <input class="input" type="text" placeholder="Enter Pressure" name="pressure">
-        </div>
-      </div>    
-      <button class="button is-primary">
-        Add Reading
-      </button>
-    </div>
-  </form>
-
-
-
-  <form class="box" action="/station/{{_id}}/autogeneratereading" method="POST">
-    <label class="label has-text-centered"><b>Auto Generate Reading</b></label>
-    <div class="field has-text-centered">   
-      <button class="button is-danger is-centered">
-        Add Auto Reading
-      </button>
-    </div>
-  </form>
+          <div class="field">
+            <input class="input" type="text" placeholder="Enter Temperature" name="temperature">
+          </div>
+          <div class="field">
+            <input class="input" type="text" placeholder="Enter Wind Speed" name="windspeed">
+          </div>
+          <div class="field">
+            <input class="input" type="text" placeholder="Enter Wind Direction" name="winddirection">
+          </div>
+          <div class="field">
+            <input class="input" type="text" placeholder="Enter Pressure" name="pressure">
+          </div>
+        </div>    
+        <button class="button is-primary">
+          Add Reading
+        </button>
+      </div>
+    </form>
+  </div>
 {{/each}}
