+++ views/partials/add-station.hbs
@@ -1,3 +1,4 @@
+<br>
 <form action="/dashboard/addstation" method="POST">
   <label class="label has-text-centered"><b>Add New Station</b></label><br>
   <div class="columns">
