+++ views/partials/data-overview.hbs
@@ -1,6 +1,6 @@
 <div class="columns">
   <div class="column">
-    <div class="box has-background-black-ter has-text-white">
+    <div class="box has-background-black has-text-white">
       <div class="columns">
         <div class="column">
           <h5 class="title is-5 has-text-warning">
@@ -20,10 +20,10 @@
     </div>
   </div>
   <div class="column">
-    <div class="box has-background-black-ter has-text-white">
+    <div class="box has-background-black has-text-white">
       <div class="columns">
         <div class="column">
-          <h5 class="title is-5 has-text-warning">
+          <h5 class="title is-5 weather">
             Weather
           </h5>
           <p>
@@ -39,10 +39,10 @@
     </div>
   </div>
   <div class="column">
-    <div class="box has-background-black-ter has-text-white">
+    <div class="box has-background-black has-text-white">
       <div class="columns">
         <div class="column">
-          <h5 class="title is-5 has-text-warning">
+          <h5 class="title is-5 temperature">
             Temperature
           </h5>
              <div class="columns">
@@ -80,10 +80,10 @@
     </div>
   </div>
   <div class="column">
-    <div class="box has-background-black-ter has-text-white">
+    <div class="box has-background-black has-text-white">
       <div class="columns ">
         <div class="column">
-          <h5 class="title is-5 has-text-warning">
+          <h5 class="title is-5 wind">
             Wind
           </h5>
              <div class="columns">
@@ -121,10 +121,10 @@
     </div>
   </div>
   <div class="column">
-    <div class="box has-background-black-ter has-text-white">
+    <div class="box has-background-black has-text-white">
       <div class="columns">
         <div class="column">
-          <h5 class="title is-5 has-text-warning">
+          <h5 class="title is-5 pressure">
             Pressure
           </h5>
           <div class="columns">
