+++ views/partials/list-readings.hbs
@@ -1,15 +1,13 @@
-<table class="table is-fullwidth">
-  <label class="label has-text-centered"><b>Weather Readings</b></label>
+<table class="table is-fullwidth is-hoverable">
   <thead>
     <tr>      
-      <th>Date & Time</th>
-      <th>Weather Code</th>
-      <th>Temperature</th>
-      <th>Wind Speed</th>
-      <th>Wind Direction</th>
-      <th>Pressure</th>
-      <th></th>
-      <!-- Adding space -->
+      <th>Date and Time</th>
+      <th><p class="weather">Weather Code</p></th>
+      <th><p class="temperature">Temperature</p></th>
+      <th><p class="wind">Wind Speed</p></th>
+      <th><p class="wind">Wind Direction</p></th>
+      <th><p class="pressure">Pressure</p></th>
+      <th><!-- Adding space --></th>
       <th>Delete</th>
     </tr>
   </thead>
