+++ views/partials/station-map.hbs
@@ -1,19 +1,21 @@
-<div id="map">
-  <script>
-    let markers = [
-      { name: "Hiroshima, Japan", lat: 34.39, long: 132.46 },
-      { name: "Longyearbyen, Svalbard", lat: 78.22, long: 15.63 },
-      { name: "Pyongyang, Democratic People's Republic of Korea", lat: 39.0738, long: 125.8198 }
-    ]
-    let map = L.map('map').setView([38, 130], 3);
-    L.tileLayer('https://tile.openstreetmap.org/{z}/{x}/{y}.png', {
-    minZoom: 2,
-    maxZoom: 11,
-    attribution: '© OpenStreetMap'
-    }).addTo(map);
-    
-    for (let i = 0; i < markers.length; i++){
-      var marker = L.marker([markers[i].lat, markers[i].long]).bindPopup(markers[i].name).addTo(map);
-    }       
-  </script>
+<div class="box">
+  <div id="map">
+    <script>
+      let markers = [
+        { name: "Hiroshima, Japan", lat: 34.39, long: 132.46 },
+        { name: "Longyearbyen, Svalbard", lat: 78.22, long: 15.63 },
+        { name: "Pyongyang, Democratic People's Republic of Korea", lat: 39.0738, long: 125.8198 }
+      ]
+      let map = L.map('map').setView([38, 130], 3);
+      L.tileLayer('https://tile.openstreetmap.org/{z}/{x}/{y}.png', {
+      minZoom: 2,
+      maxZoom: 11,
+      attribution: '© OpenStreetMap'
+      }).addTo(map);
+
+      for (let i = 0; i < markers.length; i++){
+        let marker = L.marker([markers[i].lat, markers[i].long]).bindPopup(markers[i].name).addTo(map);
+      }       
+    </script>
+  </div>
 </div>
