+++ views/partials/welcome-menu.hbs
@@ -1,4 +1,4 @@
-<nav class="navbar is-link">
+<nav class="navbar has-background-link-dark">
   <div class="navbar-brand">
     {{> brand}}
     <div class="navbar-item">
