+++ views/station-view.hbs
@@ -7,7 +7,9 @@
   {{#each station}}
     {{> data-overview}}
   {{/each}}
-  {{>station-trends}}
   {{> list-readings}}
   {{> add-reading}}
+  <br>
+  {{>station-trends}}
+  
 </section>
